toolchain: update to 1.72.0 (#9460)
The code changes largely just address warnings from the compiler. The `drop` one is fairly straightforward in that `Box::from_raw` got a `#[must_use]` annotation, and the `resolver = "2"` one was a solution to a complaint from cargo about the fact that the default resolver in `edition = "2021"` crates is version 2, but this is only applicable for individual crates and workspaces must opt-in manually.