fix(nayduck): Fix many state sync tests (#9496)
* fix(nayduck): Make many tests pass again.

The issue was misconfiguration of tests, particularly of state sync.
It's not enough to declare `state_sync_enabled: True`, the only mechanism of state sync that is officially supported requires a state dump.

In a few instances reduced the number of shards in a test, because the simpler the test the better.

* Format

* fix

* Added a state_sync library to generate configs

* Added a state_sync library to generate configs

* Refactor the epoch_height() function

* Make epochs a bit longer

* Attempt to fix gc_sync_after_sync.
Fix test_mock_node_setup.
Add debug info to test_make_state_snapshot()

* .

* .

* .

* Deflake test_mock_node_basic

* Deflake test_mock_node_basic

---------

Co-authored-by: near-bulldozer[bot] <73298989+near-bulldozer[bot]@users.noreply.github.com>