V2 Routing: Refresh distance vector nonces (#9651)
DistanceVector messages contain timestamped edges, which by default
expire after 30 minutes. However, transmission of DistanceVectors is
triggered only by changes to the network topology. As a result, nodes in
a stable network will needlessly expire their DistanceVectors after 30
minutes. When a DistanceVector expires, the connection with the
associated peer is terminated, resulting in unnecessary churn of network
connections.

RoutingTableUpdate is an existing message type used by the V1 protocol
to periodically flood refreshed edge nonces to the whole network. This
PR passes the refreshed nonces from RoutingTableUpdate messages into the
local storage for the V2 protocol so that the timestamps for stable
DistanceVectors can be kept up to date.