jsonrpc: allow parse object at tx status requests (#9648)
Adding the possibility to pass object as params for tx status methods:
```
{
  "jsonrpc": "2.0",
  "id": "dontcare",
  "method": "EXPERIMENTAL_tx_status", // could be "tx"
  "params": {
    "tx_hash": "6zgh2u9DqHHiXzdy9ouTP7oGky2T4nugqzqt9wJZwNFm",
    "sender_account_id": "sender.testnet"
  }
}
```

This is just a nice improvement, it looks cleaner.
But I specifically need this as a prep step for
https://github.com/near/nearcore/issues/6837
Having json params will allow me to add new request parameter easier.

I thought about adding passing object to send tx methods as well, but
the user has to pack it by borsh to be able to sign it.
So, deserialising SignedTransaction by serde makes no sense, I dropped
this functionality.
Serialisation can still be useful to be able to dump tx to json so that
humans may read it.