refactor(state-sync): don't make the caller of StateSync::run() expli… (#9655)
…citly save state

The caller of StateSync::run() is expected to save the HashMap passed to
this function if StateSync::run() returns StateSyncResult::Changed. This
means that StateSync::run() needs to keep track of whether it changed
anything in this hashmap throughout all the functions it calls, and then
have the caller replace its stored hash map with the cloned one passed
to this function. This is like implementing a function fn f(x: &mut u32)
{} with fn(x: &mut u32) -> bool {} where the caller is expected to clone
the argument and save it if the function returns true, which kind of
complicates the functions involved and makes refactoring other things a
little messier. It is also kind of error prone:
https://github.com/near/nearcore/pull/6939