[Trie] Optimize runtime trie access by avoiding wasted lookups of inlined values. (#10100)
Previously in #9991 , `Trie::get` was improved so that it would not
issue an on-disk lookup of a value if the value was fetched from flat
storage and was already inlined in flat storage. However, during
contract execution, the `Trie::get` code path is never executed; instead
we still always go through `Trie::get_ref`. The problem is that
`Trie::get_ref` always throws away the value even if it's already
available via `FlatStateValue::Inlined`, so that means during contract
execution, the flat state inlined value is useless.

This PR applies the same optimization for contract executions. In fact,
it cleans up the code so that it's less confusing than the state that
the previous PR left it in. We're pretty much back to `get_ref` and
`get`, except that `get_ref` is now changed to `get_optimized_ref`, and
returns a reference that can be dereferenced efficiently if the value
was already available. Gas accounting is now simpler too, because the
value is considered accessed when and only when dereferencing it.