fix: use cargo-bolero with a fix for `cargo bolero list` with failing tests (#10131)
We have tests that fail with `cargo test`, and this throws off `cargo
bolero list`, that runs `cargo test` to list the fuzzers.

With this change, we add camshaft/bolero#195, that ignores such failures
and actually lists all fuzzers.

This should fix the issue, that all the latest uploads to the fuzzers
bucket actually were empty tarballs.

---------

Co-authored-by: Simonas Kazlauskas <github@kazlauskas.me>