[resharding] GC trie state and flat state after resharding (#10137)
This PR plugs into the Garbage Collection (GC) code to cleanup trie
state and flat state after a resharding event.

During GC, when we are clearing the last block of the epoch when
resharding happens, we can also go ahead and clean up state from the
parent shards.

More context on GC:
https://near.github.io/nearcore/architecture/how/gc.html

We use the functions introduced in PR
https://github.com/near/nearcore/pull/10136 to clear the trie cache as
well as delete the flat storage and set flat storage status as empty.