fix: test_check_unknown_tx_must_return_error nayduck test (#10134)
This PR tries to fix
[test_check_unknown_tx_must_return_error](https://github.com/near/nearcore/blob/0e0f191ea31c173e266530505cf7147b14f509ab/integration-tests/src/tests/nearcore/rpc_nodes.rs#L420)
Master code: 
The timeout() returns Elapsed if after polling there are no result, and
if this happens, we return a TIMEOUT error instead of the specific error
that we got that was not returned as a result. For example, when polled
result matched `Err(err @
near_jsonrpc_primitives::types::transactions::RpcTransactionError::UnknownTransaction
{
                        ..
})`, there are some edge cases that will always be TIMEOUT; and if the
transaction never got finalized within timeout, then the error returned
will always be TIMEOUT too.

The change:
use a mutable variable to represent the result we got inside the timeout
loop. If this was indeed an error after timeout has passed, we return
this error; if it is not, that means finality is not reached, we return
TIMEOUT error.