refactor(resharding): remove the rust feature gating resharding (#10153)
As per the discussion on
[zulip](https://near.zulipchat.com/#narrow/stream/295302-general/topic/gating.20new.20features)
we no longer require the rust feature to gate new features. Removing it
will make development and mocknet testing in particular easier.

This is not stabilization and rollout, resharding should still remain in
nightly and not get enabled by default. Dear reviewers please watch out
for any place where this is not done properly.