[garbage_colection] Fix crash in clear_resharding_data while fetching prev_epoch_id (#10195)
We had an issue in the `clear_resharding_data` while fetching the
`prev_epoch_id`. The implementation of `prev_epoch_id` relied on
fetching the block_info of the last block of prev_epoch to get the
epoch_id. This unfortunately failed for the case of GC as the block_info
was already garbage collected.

The new implementation here relies on using the block_header to get the
epoch_id instead of block_info.

This was unfortunately only caught in mocknet and not integration tests
as having a small enough epoch_length lead to the block_info being
cached in the epoch_manager (even though it was GC'd)

Zulip post:
https://near.zulipchat.com/#narrow/stream/295558-pagoda.2Fcore/topic/Master.20binary.20Can't.20clear.20old.20data/near/402240517

Pending: Figure out a way to include test from this PR:
https://github.com/near/nearcore/pull/10193
We would need to enable no_cache feature for the test.