introduce code coverage analysis in CI (#10180)
Note: I’m removing the preparatory "Build neard" job here. The reason is
twofold:
- I didn’t manage to get coverage output to actually write *.profraw
files without the whole target folder
- Uploading the whole compressed (to 3G) target folder as an artifact
takes 6 minutes for compression and 15 minutes for upload
- Just building, on the commit on which I was, took 2 minutes,
multiplied by 4 it gives 8 minutes, so quite a lot less than the
compress+upload time
- So overall we’re spending less time by just redoing the compilation
work, as opposed to uploading+downloading it; and the tests themselves
are relatively fast too so having them on the bigger machines (8cores,
enough for them to not be the bottleneck) shouldn’t hog them for too
long.