reintroduce clear_resharding_data (#10211)
Tested on updated #10193

`is_next_block_epoch_start` reads `BlockInfo` of the first block of the
epoch, which is not correct in the case of garbage collection.
The previous version of the test was passing, because that `BlockInfo`
was still in the cache of the epoch manager (as we call
`is_next_block_epoch_start` on every block)

Implemented safer version of `is_next_block_epoch_start` --
`is_last_block_in_finished_epoch` (didn't think too much about the name,
open to suggestions).
`is_last_block_in_finished_epoch` works by relying on the fact that if
we processed block, and it was the last block of an epoch according to
`is_next_block_epoch_start`, then we wrote an `EpochInfo` with the hash
of that block, and `EpochInfo` is not garbage collectible.