feat(epoch-sync): EpochSyncInfo processing (#10226)
#10031 
This PR is a step towards `EpochSyncInfo` validation that doesn't
actually include any validation.
I want to combine creation of necessary `StoreUpdate` with validation,
because to validate `EpochSyncInfo` we need to rebuild `BlockMerkleTree`
for every block in the epoch, but also not commit it if anything goes
bad, and that is supported in `ChainStoreUpdate`.
Creation of `StoreUpdate` is already a big change, so I want to do a
separate PR for it.
To test that we process `EpochSyncInfo` correctly I create a VERY hacky
test, that should not be included in CI ever. Right now it works as
intended, but if we change implementation of any sync, we will have to
rewrite this test.
I hope I will add epoch sync into `run_sync_step` flow soon enough for
us to replace this test with something stable before we need to rewrite
unstable test.