refactor: reduce places where we assume shard ids are contiguous  (#10230)
There are many places in the repo where we assume that the valid shard
ids are in the range [0, num_shards). This PR is an attempt to improve
the current state of affairs.

- ShardLayout introduces a new method: `fn shard_ids()` which still
employs the above assumption.
- All instances above assumption are moved to calls to the above
function so that the assumption is centralised in a single place.

Future work:
- If we have a function that returns a list of shard ids, we do not need
`fn num_shards()`. It can be derived from the previous function. It
should be removed for consistency reasons.