coverage: apply `-Cinstrument-coverage` to workspace crates only (#10239)
When collecting coverage we want to instrument only the crates in our
workspace, and not the external dependencies. There is no value in doing
otherwise – we don’t run tests for our dependencies anyway. Furthermore,
instrumenting the external dependencies runs a risk of significant
performance regressions, such as seen in #10201.